moved the logic to archive the file back to sftp-integration-base
