+++ sources/packages/backend/libs/integrations/src/esdc-integration/fed-restriction-integration/fed-restriction.processing.service.ts
@@ -3,10 +3,7 @@ import { Injectable } from "@nestjs/common";
 import { FedRestrictionIntegrationService } from "./fed-restriction.integration.service";
 import { DataSource, InsertResult } from "typeorm";
 import { FederalRestriction, Restriction } from "@sims/sims-db";
-import {
-  getFileNameAsExtendedCurrentTimestamp,
-  getISODateOnlyString,
-} from "@sims/utilities";
+import { getISODateOnlyString } from "@sims/utilities";
 import { FedRestrictionFileRecord } from "./fed-restriction-files/fed-restriction-file-record";
 import { ProcessSFTPResponseResult } from "../models/esdc-integration.model";
 import { ConfigService, ESDCIntegrationConfig } from "@sims/utilities/config";
@@ -17,7 +14,6 @@ import {
 } from "@sims/integrations/services";
 import { SystemUsersService } from "@sims/services/system-users";
 import { StudentRestrictionSharedService } from "@sims/services";
-import * as path from "path";
 import { SFTP_ARCHIVE_DIRECTORY } from "@sims/integrations/constants";
 
 /**
@@ -81,10 +77,9 @@ export class FedRestrictionProcessingService {
       // Only the most updated file matters because it represents the entire data snapshot.
       for (const remoteFilePath of filePaths) {
         try {
-          const newRemoteFilePath = this.getArchiveFilePath(remoteFilePath);
-          await this.integrationService.renameFile(
+          await this.integrationService.archiveFile(
             remoteFilePath,
-            newRemoteFilePath,
+            SFTP_ARCHIVE_DIRECTORY,
           );
         } catch (error) {
           result.errorsSummary.push(
@@ -101,23 +96,6 @@ export class FedRestrictionProcessingService {
     }
 
     return result;
-  }
-
-  /**
-   * Gets a new file path to archive the file.
-   * @param remoteFilePath full file path with a file name.
-   * @returns new full file path with a file name.
-   */
-  private getArchiveFilePath(remoteFilePath: string) {
-    const fileInfo = path.parse(remoteFilePath);
-    const timestamp = getFileNameAsExtendedCurrentTimestamp();
-    const fileBaseName = `${fileInfo.name}_${timestamp}${fileInfo.ext}`;
-    const newRemoteFilePath = path.join(
-      fileInfo.dir,
-      SFTP_ARCHIVE_DIRECTORY,
-      fileBaseName,
-    );
-    return newRemoteFilePath;
   }
 
   /**
